requirements.txt
Django==1.5.2
Markdown==2.3.1
South==0.8.2
django-filter==0.7
django-guardian==1.1.1
djangorestframework==2.3.8
ipython==1.0.0
python-slugify==0.0.6
